fix: Use build tags for cross-platform signal handling
- Split signal handling into platform-specific files
- signals_unix.go: Full SIGWINCH support for Unix platforms
- signals_windows.go: No-op implementation for Windows
- Ensures clean cross-compilation for all target platforms
