+++ main_helpers.go
@@ -474,28 +474,3 @@ func handleInputWithTerminalState(stdinPipe io.WriteCloser, done chan bool, logg
 	}
 }
 
-// handleSignalsAndResizeWithTerminalState handles signals and window resizing with terminal state
-func handleSignalsAndResizeWithTerminalState(session *ssh.Session, termState *TerminalStateManager, logger *log.Logger) {
-	// Set up signal channel for SIGWINCH (window resize)
-	// SIGWINCH is not available on Windows, so skip signal handling on Windows
-	if runtime.GOOS == "windows" {
-		// On Windows, we don't have SIGWINCH, so just return
-		return
-	}
-	
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, syscall.SIGWINCH)
-	
-	for {
-		select {
-		case <-sigCh:
-			if termState.IsRaw() {
-				fd := termState.GetFD()
-				if width, height, err := term.GetSize(fd); err == nil {
-					// Send window size change to remote
-					session.WindowChange(height, width)
-				}
-			}
-		}
-	}
-}
